Add new cpEntry for modified anonClassName
Currently, when an anonClass is created, the entry in the constantPool
that contains the className utf is modified and is allocated a larger
buffer for the name (which re-written during ramClass creation -
className/addressOfRomClass). A problem arises when there are multiple
references to the utf8 (ie. Constant_Class ref and a Constant_String
ref) in the constantPool. If an anonClass happens to have a string
literal that is the same as its class name, this string literal gets
incorrectly modified because there is only one shared reference to the
utf in the constantPool. 

This patch creates a new constantPool entry for the className if we are
creating an anonClass and it is referenced by both a Class ref and a
String or nameAndType ref. This way all other references point to the
original utf8 but the classref points to the modified utf entry. If the
constantPool size is at U_16_MAX then we fail to load the class.

Signed-off-by: tajila <atobia@ca.ibm.com>