Add support for function pointers in Panama
When invoking a native function (with Panama) that takes in a function
pointer (ie. qsort), the groveller generates a java abstract method that
takes in a functional interface with java carrier types that match the
native types for the function.

In the native binding stage the 'generateFunctionPointerBinding'
generates a native function (via JITBuilder) that takes all the
arguments and places them in a jvalue array and calls into the jvm with
env->call[retType]methodA(...) and then returns.

Signed-off-by: Tobi Ajila <tobi_ajila@ca.ibm.com>