Add initial implementation for the Valhalla L-World withfield bytecode.
This implementation accepts a reference to a heap-allocated value class
instance from the stack and clones it on the heap. The desired field of
the clone is mutated and a reference to the new object is left on the
stack. The implementation details are similar to those of
inlInternalsPrimitiveClone() and putfield() in BytecodeInterpreter.hpp.

IncompatibleClassChangeError is thrown if the class of the instance is
not a value class.

Signed-off-by: Eric Zhang <eric99.zhang@gmail.com>