Load classes of flattenable fields during loading of declaring class
As per http://cr.openjdk.java.net/~fparain/L-world/L-World-JVMS-4b.pdf,
the class specified in a flattenable field's signature must be loaded
during the loading phase of the field's declaring class.

The existing `verifyClassLoadingStack()` function in
`createramclass.cpp` is relied on to throw `ClassCircularityError` when
a class contains itself directly or indirectly through a chain of
flattenable fields.

When loading a class, loading RAM classes of flattenable instance fields
also gives the field's instance size, which is necessary in computing
the instance size and description of the class currently being loaded.

Signed-off-by: Eric Zhang <eric99.zhang@gmail.com>

# Conflicts:
#	runtime/vm/createramclass.cpp