Add flattenable value type field support to getfield
Add support for flattenable value fields to getfield. When used with a
flattenable field, getfield creates a new instance of the value on the
heap and copies the fields of the flattened value to the new instance.

This operation requires the J9Class of the field, which is not available
from the J9RAMFieldRef. This version uses internalFindClassUTF8 to
locate the J9Class on every use of getfield with a flattened field. A
subsequent version could, in the parent J9Class, keep an array of
J9Classes of flattenable fields, and then use 16 bits of the
J9RAMFieldRef's valueOffset as an index to the correct J9Class.

Signed-off-by: Eric Zhang <eric99.zhang@gmail.com>