Add flattenable value type field support to putfield
Add support for flattenable value fields to putfield. When used with a
flattenable field, putfield copies the fields of the given value into
the flattened value field. If the value is null, a NullPointerException
is thrown.

This operation requires the J9Class of the field, which is not available
from the J9RAMFieldRef. This version uses internalFindClassUTF8 to
locate the J9Class on every use of putfield with a flattened field. A
subsequent version could, in the parent J9Class, keep an array of
J9Classes of flattenable fields, and then use 16 bits of the
J9RAMFieldRef's valueOffset as an index to the correct J9Class.

Signed-off-by: Eric Zhang <eric99.zhang@gmail.com>