Add flattening support for valuetypes
This PR does not address volatile access of flattened fields
This PR does not address flattened value hidden fields
This PR does not allow flattened value contended fields
This PR does not address flattened value static fields
This PR does not address flattened value arrays

This PR introduces an option to -XX:ValueTypeFlatteningThreshold 
to set the threshold at which a flattenable value type is no 
longer flattened. 

Fields are ordered in the following manner:
- FlattenedTypeContainingDoubles
- Doubles
- FlattenedTypeContainingObjects
- Objects
- FlattenedTypeContainingSingles
- Singles

A valueType with a Q signature is flattenable. However, the JVM 
will only flatten it if the J9ClassIsFlattened bit is set in the 
classFlags (this is determined by the flattening threshold). If 
the instance field is resolved, flattened fields will have the 
J9FieldFlagFlattened bit set in the ram CP flags slot. If this
bit is set, the offset contained in the ram CP entry is not the 
field offset, rather it is an offset into the 
clazz->flattenedFieldCache[] table. Each entry in this table 
contains the J9Class for the flattened field and the real field 
offset.

When a field is flattened it does not contain a the lockword only 
instance fields. In order to do anything with a field one must
first retrieve it with the getField bytecode, this bytecode will 
create a new object and copy the contents of the flattened field 
into the new object, then place a reference to the new object on 
the stack. The new object will have a lockword.

Some of this work is based on the PackedObjects tech preview and 
the work done by Eric Zhang.

Signed-off-by: tajila <atobia@ca.ibm.com>