Ram cache prototype
These are the ramClass cache prototype changes. It works with
non-compressedrefs 64 bit JVMs with ASLR turned off.

The original design is here:
https://ibm.box.com/s/973c953e9sc50xrj00u7y43nj6a6892l

Option to run it is:

./java -Xshare:off -Xsnapshot=cache -version

The first run will create the snapshot the subsequent runs will use it.

Its been fixed up a bit from the original repo. Some of the changes are:

- Removed the call-ins to rebuild class objects since we will be
persisting the java heap. As a result Quarkus will not work but small
simple apps will.
- Renamed cold to snapshot, warm to restore.
- Renamed -Xramcache to -Xsnapshot

Issues have been created to address the shortcomings of the prototype.

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>