Persist portlibrary in image file
Even with ASLR disabled shared libs files can be mapped into different
memory addresses from run to run. This can happen if there are
differences (difference from snapshot to restore run) in the way java
launcher (java.c) allocates memory before it calls createJavaVM. Any
difference can cause the libj9vm.so to be loaded at a different address.
I have been able to observe this consistently by simply varying the
length of cmdline args (which is necessary since snapshot args will be
longer than restore).

THe problem with shared libs moving around is that there are persisted
structures which have pointers to code, if the address of code changes
then these pointers need to be updated.

This PR specifically addresses the cases where structures (such as
hashtables) have direct references to the portlib which has pointer to
code. This PR persists the portlib so that on a restore run only the
portlib has to be fixed up and not all the places that referred to it.

This PR also removes the MAP_FIXED arg to mmap and verifies that the
mapped address is what is expected, if not JVM startup fails. I am yet
to observe a startup failure due to this with ASLR turned off.

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>