Ensure JDWP threads are not halted when hooks run for CRIU
A breakpoint set during a pre-checkpoint or post-restore hook will cause Java
to hang because the JDWP threads are suspended at that time. This change will
delay the suspending of JDWP threads for pre-checkpoint hooks and resume them
early for post-restore hooks if VM is run with the JDWP agentlib loaded.

Signed-off-by: Mike Zhang <mike.h.zhang@ibm.com>