Add support for FFI_TYPE_STRUCT_FF/DD on z/OS
XPLINK on z/OS requires a special treatment for structs which with two
members whose base type are same, either float or double. Such
parameters are passed to the function as complex type and available FPRs
will be used.
Changes in this PR adds support for two new types in libffi -
FFI_TYPE_STRUCT_FF and FFI_TYPE_STRUCT_DD and also uses it in J9 similar
to how FFI_TYPE_STRUCT is used as for J9 it shouldn't be different. It
is just that libffi needs to know if the argument type or return type
are considered complex on XPLINK to make sure that appropriate FPRs if
available are used to pass parameters or read return values from.

J9 does not need to do the analysis for complex types, that work is done
when platform dependant prep_cif is called when initializing the cif
object.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>