Make AOTCacheRecord::setSubrecordPointers protected
This works around a bug in clang. Without this, the invocation of
setSubrecordPointers in the AOTCacheRecord::readRecord<> template method
fails; clang complains that setSubrecordPointers is a private method of
AOTCacheRecord. This shouldn't matter at all, as a method of
AOTCacheRecord ought to be able to call the private methods of the
AOTCacheRecord class itself, but clang issues an erroneous access error
anyway.

Signed-off-by: Christian Despres <despresc@ibm.com>