Send cached J9ROMClass hashes to JITServer
If a JITServer client has a cached hash for a ROMClass, it may now send
this hash to the server to avoid the server having to re-hash (and
possibly re-pack) the ROMClass received from the client. The server will
store this cached hash in its shared ROMClass cache. Currently, only the
stored hashes of generated classes are sent to the server.

Signed-off-by: Christian Despres <despresc@ibm.com>