Cap R0's value to 3 when handling complex structs in sysvz64.s
R0's value is used to calculate a branch offset. This value can be
between 0 and 3. Due to the nature of how complex types are passed in
registers, it's possible to increment this value beyond 3. This commit adds
some support code to cap that value at 3 before any branch offsets are computed.